zapata_m3/requirements.txt
python-dotenv
chromadb
openai
transformers
pdfminer.six
pdfplumber
customtkinter
pandas
scikit-learn
tqdm
psutil
requests
colorlog
rapidfuzz
spacy
tensorflow_hub
sentence_transformers
rapidfuzz
